update file locations
