+++ src/frontend/screens/MapScreen/MapViewMemoized.tsx
@@ -2,10 +2,16 @@ import * as React from 'react';
 import Mapbox, {UserLocation} from '@rnmapbox/maps';
 import config from '../../../config.json';
 import CheapRuler from 'cheap-ruler';
-import {IconButton} from '../IconButton';
-import {LocationFollowingIcon, LocationNoFollowIcon} from '../icons';
+import {IconButton} from '../../sharedComponents/IconButton';
+import {
+  LocationFollowingIcon,
+  LocationNoFollowIcon,
+} from '../../sharedComponents/icons';
 import {View, StyleSheet} from 'react-native';
 import {ObservationMapLayer} from './ObsevationMapLayer';
+import {AddButton} from '../../sharedComponents/AddButton';
+import {useNavigationFromHomeTabs} from '../../hooks/useNavigationWithTypes';
+import {useDraftObservation} from '../../hooks/useDraftObservation';
 
 // This is the default zoom used when the map first loads, and also the zoom
 // that the map will zoom to if the user clicks the "Locate" button and the
@@ -25,10 +31,17 @@ type MapViewProps = {
 
 const MAP_STYLE = Mapbox.StyleURL.Outdoors;
 
-export const MapView = React.memo(
+export const MapViewMemoized = React.memo(
   ({coords, locationServiceEnabled, isFocused}: MapViewProps) => {
     const [zoom, setZoom] = React.useState(DEFAULT_ZOOM);
     const [following, setFollowing] = React.useState(false);
+    const {newDraft} = useDraftObservation();
+    const {navigate} = useNavigationFromHomeTabs();
+
+    const handleAddPress = () => {
+      newDraft();
+      navigate('CategoryChooser');
+    };
 
     React.useEffect(() => {
       Mapbox.setTelemetryEnabled(false);
@@ -40,7 +53,7 @@ export const MapView = React.memo(
     }
 
     return (
-      <React.Fragment>
+      <View style={{flex: 1}}>
         <Mapbox.MapView
           testID="mapboxMapView"
           style={{flex: 1}}
@@ -82,7 +95,8 @@ export const MapView = React.memo(
             </IconButton>
           </View>
         )}
-      </React.Fragment>
+        <AddButton testID="addButtonMap" onPress={handleAddPress} />
+      </View>
     );
   },
   shouldComponentSkipRerender,
